fake-news-detector/requirements.txt
streamlit
python-dotenv
requests
beautifulsoup4
pytest
black
flake8